requirements.txt
Flask==2.0.2
torch==1.9.0  
Pillow==8.3.2
ultralytics==8.0.0 
